refactor/#104: app-jwt.enum.ts 파일명 수정
