+++ src/features/auth/commands/generate-access-token/generate-access-token.command-handler.ts
@@ -3,7 +3,7 @@ import { UserRepositoryPort } from '@features/user/repositories/user.repository-
 import { USER_REPOSITORY_DI_TOKEN } from '@features/user/tokens/di.token';
 import { AppJwtServicePort } from '@libs/app-jwt/services/app-jwt.service-port';
 import { APP_JWT_SERVICE_DI_TOKEN } from '@libs/app-jwt/tokens/app-jwt.di-token';
-import { TokenType } from '@libs/app-jwt/types/jwt.enum';
+import { TokenType } from '@libs/app-jwt/types/app-jwt.enum';
 import { HttpUnauthorizedException } from '@libs/exceptions/client-errors/exceptions/http-unauthorized.exception';
 import { COMMON_ERROR_CODE } from '@libs/exceptions/types/errors/common/common-error-code.constant';
 import { isNil } from '@libs/utils/util';
